refactor: setup next alias automatically
