+++ packages/react-server/src/next/plugin.ts
@@ -7,15 +7,32 @@ import {
   vitePluginLogger,
   vitePluginSsrMiddleware,
 } from "@hiogawa/vite-plugin-ssr-middleware";
-import type { PluginOption } from "vite";
+import type { AliasOptions, Plugin, PluginOption } from "vite";
+
+const nextAlias: AliasOptions = [
+  {
+    find: /^next(\/.*)?/,
+    replacement: "@hiogawa/react-server/next/compat$1",
+  },
+];
+
+const nextAliasPlugin: Plugin = {
+  name: "next-compat-alias",
+  config: () => ({
+    resolve: {
+      alias: nextAlias,
+    },
+  }),
+};
 
 export function vitePluginReactServerNext(options?: {
   plugins?: PluginOption[];
 }): PluginOption {
   return [
+    nextAliasPlugin,
     vitePluginReactServer({
       routeDir: "app",
-      plugins: options?.plugins,
+      plugins: [nextAliasPlugin, ...(options?.plugins ?? [])],
     }),
     vitePluginLogger(),
     vitePluginSsrMiddleware({
